+++ docs/changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
+### 2025-09-08
+
+- feat: add terraform module
+
 ### 2025-09-05
+
 - fix: issue with agent ready file not created on setup failure, leading to blockage of further
     operations.
 - docs: Fixed Vale and link check failures.
@@ -11,5 +16,6 @@
     operations.
 
 ### 2025-03-11
+
 - docs: added changelog
 - chore: removed src-docs